+++ sw-sysemu/memory/maxion_region.h
@@ -10,7 +10,7 @@
 #include "memory_error.h"
 #include "memory_region.h"
 
-extern uint32_t current_thread;
+extern unsigned current_thread;
 
 namespace bemu {
 
@@ -18,7 +18,7 @@ namespace bemu {
 extern typename MemoryRegion::value_type memory_reset_value;
 
 
-template<unsigned long long Base, size_t N>
+template<unsigned long long Base, unsigned long long N>
 struct MaxionRegion : public MemoryRegion {
     typedef typename MemoryRegion::addr_type      addr_type;
     typedef typename MemoryRegion::size_type      size_type;
