+++ sw-sysemu/memory/sparse_region.h
@@ -16,7 +16,8 @@ namespace bemu {
 extern typename MemoryRegion::value_type memory_reset_value;
 
 
-template <unsigned long long Base, size_t N, size_t M, bool Writeable=true>
+template <unsigned long long Base, unsigned long long N, unsigned long long M,
+          bool Writeable=true>
 struct SparseRegion : public MemoryRegion
 {
     typedef typename MemoryRegion::addr_type      addr_type;
